86 add error demo
