+++ src/Components/Select/Select.js
@@ -1,29 +1,110 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import withTheme from '../../Theme/withTheme';
-import SelectOutlined from './SelectOutlined/SelectOutlined';
-import SelectFlat from './SelectFlat/SelectFlat';
-import SelectFilled from './SelectFilled/SelectFilled';
+import { TouchableHighlight, View, FlatList } from 'react-native';
 
-class TextField extends Component {
+import { Menu, MenuItem, Icon, TextField } from '../../..';
+
+import withTheme from '../../Theme/withTheme';
+import styles from './Select.styles';
+
+class Select extends Component {
   static propTypes = {
+    onSelect: PropTypes.func,
+    buttonStyle: PropTypes.object,
+    label: PropTypes.string,
+    selectedItem: PropTypes.node,
+    menuItems: PropTypes.array,
+    visible: PropTypes.bool,
+    textFieldProps: PropTypes.object,
     type: PropTypes.string,
-    onFocus: PropTypes.func,
-    onBlur: PropTypes.func,
-    disabled: PropTypes.bool,
+    onBackdropPress: PropTypes.func,
   };
 
-  render() {
-    const { type, ...rest } = this.props;
+  state = {
+    visible: false,
+  };
 
-    if (type == 'outlined') {
-      return <SelectOutlined {...rest} />;
-    } else if (type == 'filled') {
-      return <SelectFilled {...rest} />;
-    } else {
-      return <SelectFlat {...rest} />;
-    }
+  onSelect = item => {
+    this.props.onSelect(item);
+    this.hideMenu();
+  };
+
+  showMenu() {
+    this.setState({
+      visible: true,
+    });
+  }
+
+  hideMenu() {
+    this.setState({
+      visible: false,
+    });
+  }
+
+  render() {
+    const {
+      buttonStyle,
+      label,
+      selectedItem,
+      menuItems,
+      textFieldProps,
+      type,
+    } = this.props;
+
+    const { visible } = this.state;
+    return (
+      <Menu
+        style={[styles.menu, { flex: 1 }]}
+        sameWidth
+        visible={visible}
+        modalMenuStyle={{ marginTop: 72 }}
+        onBackdropPress={() => this.hideMenu()}
+        button={
+          <TouchableHighlight
+            onPress={() => this.showMenu()}
+            style={[styles.button, buttonStyle]}
+            underlayColor={'transparent'}>
+            <View style={styles.innerView}>
+              <TextField
+                type={type ? type : 'flat'}
+                label={label}
+                value={selectedItem}
+                focused={visible}
+                editable={false}
+                style={[
+                  styles.textField,
+                  textFieldProps && textFieldProps.style,
+                ]}
+                trailingIcon={
+                  <Icon
+                    name="arrow-drop-down"
+                    size={24}
+                    color={'#757575'}
+                    style={styles.icon}
+                  />
+                }
+                {...textFieldProps}
+              />
+            </View>
+          </TouchableHighlight>
+        }>
+        <FlatList
+          data={menuItems}
+          style={{ flex: 1 }}
+          renderItem={({ item }) => {
+            return (
+              <MenuItem
+                key={item.id}
+                text={item.name}
+                onPress={() => this.onSelect(item)}
+                style={{ flex: 1 }}
+              />
+            );
+          }}
+        />
+      </Menu>
+    );
   }
 }
 
-export default withTheme(TextField);
+export default withTheme(Select);
