+++ src/Components/Select/SelectFlat/SelectFlat.stories.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { View } from 'react-native';
 
 import { Select } from '../../..';
 import { State, Store } from '@sambego/storybook-state';
@@ -24,14 +25,27 @@ export default storiesOf('Components|Select/Flat', module)
 
       <State store={store} style={{ flex: 1 }}>
         {state => (
-          <Select
-            type={'flat'}
-            label={'Select'}
-            menuItems={data}
-            onSelect={value => store.set({ selectedItem: value.name })}
-            selectedItem={state.selectedItem}
-            onBackdropPress={() => store.set({ selectedItem: false })}
-          />
+          <View>
+            <Select
+              type={'flat'}
+              label={'Select'}
+              menuItems={data}
+              onSelect={value => store.set({ selectedItem: value.name })}
+              selectedItem={state.selectedItem}
+            />
+            <Select
+              type={'flat'}
+              label={'Select'}
+              menuItems={data}
+              onSelect={value => store.set({ selectedItemTwo: value.name })}
+              selectedItem={state.selectedItemTwo}
+              textFieldProps={{
+                error: true,
+                helperText: 'error',
+                containerStyle: { marginTop: 40 },
+              }}
+            />
+          </View>
         )}
       </State>
     </Container>
